added logged-in checks to ManageAccount

+++ backend/routes/account/ManageAccount.js
@@ -6,24 +6,30 @@ const ManageAccountAPI = {
     user: Fire.auth().currentUser,
 
     logout: function() {
-        Fire.auth().signOut()
-        .then(function() { return true; })                  // user signed out successfully.
-        .catch(function(error) { return false; });          // an error happened.
+        if (this.checkLoggedIn()) {
+            Fire.auth().signOut()
+            .then(function() { return true; })                  // user signed out successfully.
+            .catch(function(error) { return false; });          // an error happened.
+        }
     },
 
     resetPassword: function() {
-        var auth = firebase.auth();
-        var emailAddress = user.email;
-
-        auth.sendPasswordResetEmail(emailAddress)
-        .then(function() { return true; })
-        .catch(function(error) { return false; });
+        if (this.checkLoggedIn()) {
+            var auth = firebase.auth();
+            var emailAddress = user.email;
+    
+            auth.sendPasswordResetEmail(emailAddress)
+            .then(function() { return true; })
+            .catch(function(error) { return false; });
+        }
     },
 
     deleteAccount: function() {
-        user.delete()
-        .then(function() { return true; })                  // user deleted successfully.
-        .catch(function(error) { return false; });          // an error happened.
+        if (this.checkLoggedIn()) {
+            user.delete()
+            .then(function() { return true; })                  // user deleted successfully.
+            .catch(function(error) { return false; });          // an error happened.
+        }
     },
 
     checkLoggedIn: function() {
